+++ cmd/imageproxy/main.go
@@ -43,7 +43,6 @@ const defaultMemorySize = 100
 
 var addr = flag.String("addr", "localhost:8080", "TCP address to listen on")
 var allowHosts = flag.String("allowHosts", "", "comma separated list of allowed remote hosts")
-var whitelist = flag.String("whitelist", "", "deprecated. use 'allowHosts' instead")
 var denyHosts = flag.String("denyHosts", "", "comma separated list of denied remote hosts")
 var referrers = flag.String("referrers", "", "comma separated list of allowed referring hosts")
 var baseURL = flag.String("baseURL", "", "default base URL for relative remote URLs")
@@ -62,10 +61,6 @@ func init() {
 
 func main() {
 	flag.Parse()
-	if *allowHosts == "" {
-		// backwards compatible with old naming of the flag
-		*allowHosts = *whitelist
-	}
 
 	p := imageproxy.NewProxy(nil, cache.Cache)
 	if *allowHosts != "" {